feat: integrating memfs
refactored the code to make use of `memfs`, which in this case removes the need to having to create a temp directory (before writing to the output directory). Instead the tool now simply uses an in-memory filesystem, and synchronizes only once at the end, reducing the file I/O operations done on the actual disk.

This should fix (hopefully) all the errors "Error: EACCES: permission denied, mkdir '.../v1.0/src$'". Additional testing in progress...